requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
google-genai>=1.0.0
pyyaml>=6.0
aiohttp>=3.9.0
tenacity>=8.2.0
python-dateutil>=2.8.0
thefuzz>=0.22.0
lxml>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
